requirements/dev.txt
-r prod.txt
pandas-stubs>=1.5.1.221024
boto3-stubs>=1.26.12
black>=22.6.0
mypy==1.3.0
mypy-extensions==1.0.0
pylint>=2.14.4
pytest>=7.1.2
sqlalchemy-stubs>=0.4